Added checking input directory if it exists.
